Add test for expired lease
