+++ extensions/azure/cosmos/cosmos-common/build.gradle.kts
@@ -14,6 +14,7 @@
 
 plugins {
     `java-library`
+    `java-test-fixtures`
 }
 
 val cosmosSdkVersion: String by project
@@ -26,9 +27,10 @@ dependencies {
     implementation("com.azure:azure-cosmos:${cosmosSdkVersion}")
     implementation("net.jodah:failsafe:${jodahFailsafeVersion}")
 
-
     testImplementation(testFixtures(project(":common:util")))
     testImplementation(testFixtures(project(":extensions:azure:azure-test")))
+
+    testFixturesImplementation("com.azure:azure-cosmos:${cosmosSdkVersion}")
 }
