+++ extensions/azure/cosmos/transfer-process-store-cosmos/build.gradle.kts
@@ -30,6 +30,7 @@ dependencies {
 
     testImplementation(testFixtures(project(":common:util")))
     testImplementation(testFixtures(project(":extensions:azure:azure-test")))
+    testImplementation(testFixtures(project(":extensions:azure:cosmos:cosmos-common")))
     testImplementation("org.awaitility:awaitility:${awaitility}")
 }
