+++ extensions/azure/cosmos/transfer-process-store-cosmos/src/test/java/org/eclipse/dataspaceconnector/transfer/store/cosmos/CosmosTransferProcessStoreIntegrationTest.java
@@ -16,10 +16,8 @@ package org.eclipse.dataspaceconnector.transfer.store.cosmos;
 
 import com.azure.cosmos.CosmosContainer;
 import com.azure.cosmos.CosmosDatabase;
-import com.azure.cosmos.CosmosScripts;
 import com.azure.cosmos.implementation.BadRequestException;
 import com.azure.cosmos.models.CosmosItemResponse;
-import com.azure.cosmos.models.CosmosStoredProcedureProperties;
 import com.azure.cosmos.models.CosmosStoredProcedureRequestOptions;
 import com.azure.cosmos.models.CosmosStoredProcedureResponse;
 import com.azure.cosmos.models.PartitionKey;
@@ -49,7 +47,6 @@ import java.time.temporal.ChronoUnit;
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
-import java.util.Scanner;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
@@ -57,6 +54,7 @@ import java.util.stream.IntStream;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.awaitility.Awaitility.await;
+import static org.eclipse.dataspaceconnector.azure.cosmos.util.StoredProcedureTestUtils.uploadStoredProcedure;
 import static org.eclipse.dataspaceconnector.transfer.store.cosmos.TestHelper.createTransferProcess;
 import static org.eclipse.dataspaceconnector.transfer.store.cosmos.TestHelper.createTransferProcessDocument;
 
@@ -87,22 +85,6 @@ class CosmosTransferProcessStoreIntegrationTest {
         if (database != null) {
             var databaseDelete = database.delete();
             assertThat(databaseDelete.getStatusCode()).isBetween(200, 400);
-        }
-    }
-
-    private static void uploadStoredProcedure(CosmosContainer container, String name) {
-        var is = Thread.currentThread().getContextClassLoader().getResourceAsStream(name + ".js");
-        if (is == null) {
-            throw new AssertionError("The input stream referring to the " + name + " file cannot be null!");
-        }
-
-        Scanner s = new Scanner(is).useDelimiter("\\A");
-        String body = s.hasNext() ? s.next() : "";
-        CosmosStoredProcedureProperties props = new CosmosStoredProcedureProperties(name, body);
-
-        CosmosScripts scripts = container.getScripts();
-        if (scripts.readAllStoredProcedures().stream().noneMatch(sp -> sp.getId().equals(name))) {
-            CosmosStoredProcedureResponse storedProcedure = scripts.createStoredProcedure(props);
         }
     }
 
@@ -185,13 +167,13 @@ class CosmosTransferProcessStoreIntegrationTest {
     @Test
     void nextForState_leaseByAnotherHolderExpired() {
         String id1 = UUID.randomUUID().toString();
-        var tp = createTransferProcess(id1, TransferProcessStates.UNSAVED);
+        var tp = createTransferProcess(id1, TransferProcessStates.INITIAL);
         TransferProcessDocument item = new TransferProcessDocument(tp, partitionKey);
         Duration leaseDuration = Duration.ofSeconds(5);
         item.acquireLease("another-connector", leaseDuration);
         container.upsertItem(item);
 
-        List<TransferProcess> processesBeforeLeaseBreak = store.nextForState(TransferProcessStates.INITIAL.code(), 2);
+        List<TransferProcess> processesBeforeLeaseBreak = store.nextForState(TransferProcessStates.INITIAL.code(), 10);
         assertThat(processesBeforeLeaseBreak).isEmpty();
 
         await()
@@ -199,7 +181,7 @@ class CosmosTransferProcessStoreIntegrationTest {
                 .pollInterval(Duration.ofSeconds(1))
                 .pollDelay(leaseDuration) //give the lease time to expire
                 .untilAsserted(() -> {
-                    List<TransferProcess> processesAfterLeaseBreak = store.nextForState(TransferProcessStates.INITIAL.code(), 2);
+                    List<TransferProcess> processesAfterLeaseBreak = store.nextForState(TransferProcessStates.INITIAL.code(), 10);
                     assertThat(processesAfterLeaseBreak).hasSize(1);
                 }
         );
